Fix read handling.
git-svn-id: file:///dev/shm/somefilter.svn@317 e102df66-1e2e-11dd-9b44-c24451a4db5e

+++ util/file_piece.hh
@@ -77,7 +77,7 @@ class FilePiece {
     const off_t total_size_;
     const off_t page_;
 
-    off_t default_map_size_;
+    size_t default_map_size_;
     off_t mapped_offset_;
 
     // Order matters: file_ should always be destroyed after this.
